requirement.txt
transformers==4.27.1
torchmetrics==0.11.0
tokenizers==0.12.1
pytorch-lightning==1.7.7
omegaconf==2.2.3
pillow==9.3.0
metis==0.2a5